+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Changed Scapegoat's balance to 0. ([#91](https://github.com/hjwylde/werewolf/issues/91))
 * Grouped `help commands` to improve readability. ([#97](https://github.com/hjwylde/werewolf/issues/97))
 * Changed the `status` and `ping` commands to tell the caller when the game is over. ([#89](https://github.com/hjwylde/werewolf/issues/89))
+* Added roles in game to the `status` command. ([#93](https://github.com/hjwylde/werewolf/issues/93))
 
 #### v0.4.7.0